+++ src/config/env.config.ts
@@ -10,7 +10,7 @@ export const envConfig = {
   hyperedgeMongoDbUri: process.env.LENSTACK_FABRIC_DB_URI,
   payMongoDbUri: process.env.LENSTACK_PAY_DB_URI,
   ledgerscanMongoDbUri: process.env.LENSTACK_LEDGERSCAN_DB_URI,
-  snowlakeMongoDbUri: process.env.LENSTACK_NFTSTUDIO_DB_URI,
+  nftstudioMongoDbUri: process.env.LENSTACK_NFTSTUDIO_DB_URI,
   swapMongoDbUri: process.env.LENSTACK_SWAP_DB_URI,
   redisSocketHost: process.env.REDIS_SOCKET,
   redisPassword: process.env.REDIS_PASSWORD,