+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Rebranding of Edgepay as Lenstack Pay
 * Rebranding of Swapstream as Lenstack Swap
 * Rebranding of Edgescan as Lenstack Ledgerscan
+* Rebranding of Snowlake as Lenstack NFT Studio
 
 ### v3.0
 * Created DTO for all controllers wherever required to avoid raw data extraction from request body